+++ R/amp_scaling.R
@@ -5,7 +5,7 @@
 #' concentration reference for proton spectroscopic imaging" by Gasparovic et al
 #' MRM 2006 55(6):1219-26. 1.5 Tesla relaxation assumptions are taken from this
 #' paper. For 3 Tesla data, relaxation assumptions are taken from "NMR 
-#' relaxation times in the human brain at 3.0 tesla" by Wansapura et al J Magn
+#' relaxation times in the human brain at 3.0 Tesla" by Wansapura et al J Magn
 #' Reson Imaging 1999 9(4):531-8. 
 #' 
 #' @param fit_result result object generated from fitting.
@@ -64,6 +64,66 @@ scale_amp_molal_pvc <- function(fit_result, ref_data, p_vols, te, tr, ...){
                                           (p_vols[["GM"]] + p_vols[["WM"]])
   
   # append tables with %GM, %WM, %CSF and %Other
+  pvc_cols <- 6:(5 + amp_cols * 2)
+  fit_result$res_tab[, pvc_cols] <- fit_result$res_tab[, pvc_cols] *
+                                    corr_factor / w_amp
+  
+  return(fit_result)
+}
+
+#' Apply water reference scaling to a fitting results object to yield metabolite 
+#' quantities in millimolar (mM) units (mol / kg of tissue water).
+#' 
+#' Note, this function assumes the volume contains a homogeneous voxel, eg pure
+#' WM, GM or  CSF. Also note that in the case of a homogeneous voxel the
+#' relative densities of MR-visible water (eg GM=0.78, WM=0.65, and CSF=0.97)
+#' cancel out and don't need to be considered. Use scale_amp_molal_pvc for
+#' volumes containing  multiple compartments. Details of this method can be
+#' found in "Use of tissue water as a concentration reference for proton
+#' spectroscopic imaging" by Gasparovic et al MRM 2006 55(6):1219-26.
+#' 
+#' @param fit_result result object generated from fitting.
+#' @param ref_data water reference MRS data object.
+#' @param te the MRS TE in seconds.
+#' @param tr the MRS TR in seconds.
+#' @param water_t1 assumed water T1 value.
+#' @param water_t2 assumed water T2 value.
+#' @param metab_t1 assumed metabolite T1 value.
+#' @param metab_t2 assumed metabolite T2 value.
+#' @param ... additional arguments to get_td_amp function.
+#' @return A \code{fit_result} object with a rescaled results table.
+#' @export
+scale_amp_molal <- function(fit_result, ref_data, te, tr, water_t1, water_t2,
+                            metab_t1, metab_t2, ...){
+  
+  if (!identical(dim(fit_result$data$data)[2:6], dim(ref_data$data)[2:6])) {
+    stop("Mismatch between fit result and reference data dimensions.")
+  }
+  
+  # check if res_tab_unscaled exists, and if not create it
+  if (is.null(fit_result$res_tab_unscaled)) {
+    fit_result$res_tab_unscaled <- fit_result$res_tab
+  } else {
+    fit_result$res_tab <- fit_result$res_tab_unscaled
+  }
+  
+  R_water   <- exp(-te / water_t2) * (1.0 - exp(-tr / water_t1))
+  R_metab   <- exp(-te / metab_t2) * (1.0 - exp(-tr / metab_t1))
+  
+  water_conc <- 55510.0
+  
+  corr_factor <- R_water / R_metab * water_conc
+  
+  amp_cols <- fit_result$amp_cols
+  
+  w_amp <- as.numeric(get_td_amp(ref_data, ...))
+  
+  if (length(w_amp) != nrow(fit_result$res_tab)) {
+    stop("Mismatch between fit result and reference data.")
+  }
+  
+  fit_result$res_tab$w_amp <- w_amp
+  
   pvc_cols <- 6:(5 + amp_cols * 2)
   fit_result$res_tab[, pvc_cols] <- fit_result$res_tab[, pvc_cols] *
                                     corr_factor / w_amp