+++ R/mrs_read_pfile.R
@@ -77,7 +77,8 @@ read_pfile <- function(fname, n_ref_scans = NULL, verbose, extra) {
   ref <- def_ref()
   nuc <- def_nuc()
   
-  meta <- list(EchoTime = hdr$te)
+  meta <- list(EchoTime = hdr$te,
+               Manufacturer = "GE")
   
   mrs_data <- mrs_data(data = data, ft = hdr$ps_mps_freq / 10, resolution = res,
                        ref = ref, nuc = nuc, freq_domain = freq_domain,
@@ -182,7 +183,7 @@ get_pfile_dict <- function(hdr_rev, con) {
     loc$ycsi        <- 440
     loc$zcsi        <- 442
     loc$ps_mps_freq <- 488
-    loc$te <- 1148
+    loc$te          <- 1148
   } else if ((floor(hdr_rev) > 11) && (floor(hdr_rev) < 25)) {
     loc$hdr_rev     <- 0
     loc$off_data    <- 1468