+++ R/rats.R
@@ -19,13 +19,25 @@
 #' @param phase_corr apply phase correction (in addition to frequency). TRUE by
 #' default.
 #' @param ret_corr_only return the corrected mrs_data object only.
+#' @param zero_freq_shift_t0 perform a linear fit to the frequency shifts and
+#' set the (linearly modeled) shift to be 0 Hz for the first dynamic scan.
+#' @param remove_freq_outliers remove dynamics based on their frequency shift.
+#' @param freq_outlier_thresh threshold to remove frequency outliers.
+#' @param remove_phase_outliers remove dynamics based on their phase shift.
+#' @param phase_outlier_thresh threshold to remove phase outliers.
+#' @param remove_amp_outliers remove dynamics based on their amplitude change.
+#' @param amp_outlier_thresh threshold to remove amplitude outliers.
 #' @return a list containing the corrected data; phase and shift values in units
 #' of degrees and Hz respectively.
 #' @export
 rats <- function(mrs_data, ref = NULL, xlim = c(4, 0.5), max_shift = 20,
                  p_deg = 2, sp_N = 2, sp_deg = 3, max_t = 0.2,
                  basis_type = "poly", rescale_output = TRUE,
-                 phase_corr = TRUE, ret_corr_only = TRUE) {
+                 phase_corr = TRUE, ret_corr_only = TRUE,
+                 zero_freq_shift_t0 = FALSE, remove_freq_outliers = FALSE,
+                 freq_outlier_thresh = 3, remove_phase_outliers = FALSE,
+                 phase_outlier_thresh = 3, remove_amp_outliers = FALSE,
+                 amp_outlier_thresh = 3) {
   
   if (inherits(mrs_data, "list")) {
     
@@ -38,7 +50,15 @@ rats <- function(mrs_data, ref = NULL, xlim = c(4, 0.5), max_shift = 20,
     res <- lapply(mrs_data, rats, ref = ref, xlim = xlim, max_shift = max_shift,
                   p_deg = p_deg, sp_N = sp_N, sp_deg = sp_deg, max_t = max_t,
                   basis_type = basis_type, rescale_output = rescale_output,
-                  phase_corr = phase_corr, ret_corr_only = ret_corr_only) 
+                  phase_corr = phase_corr, ret_corr_only = ret_corr_only,
+                  zero_freq_shift_t0 = zero_freq_shift_t0,
+                  remove_freq_outliers = remove_freq_outliers,
+                  freq_outlier_thresh = freq_outlier_thresh,
+                  remove_phase_outliers = remove_phase_outliers,
+                  phase_outlier_thresh = phase_outlier_thresh,
+                  remove_amp_outliers = remove_amp_outliers,
+                  amp_outlier_thresh = amp_outlier_thresh)
+    
     return(res)
   }
   
@@ -101,6 +121,39 @@ rats <- function(mrs_data, ref = NULL, xlim = c(4, 0.5), max_shift = 20,
   bl_spec$data <- res[,,,,,,(length(inds) + 4):(2 * length(inds) + 3),
                       drop = FALSE]
   
+  corr_dims <- dim(shifts)
+  
+  if (zero_freq_shift_t0 | remove_freq_outliers) {
+    
+    # model frequency drift
+    x <- 0:(length(shifts) - 1)
+    shift_fit <- as.numeric(stats::predict(stats::lm(as.numeric(shifts) ~ x)))
+    
+    if (zero_freq_shift_t0) shifts <- shifts - shift_fit[1]
+    
+    if (remove_freq_outliers) {
+      res <- as.numeric(shifts) - shift_fit
+      res_med <- stats::median(res)
+      res_mad <- stats::mad(res)
+      upper_lim_freq <- res_mad * freq_outlier_thresh
+      bad_freqs <- Mod(res - res_med) > upper_lim_freq
+    }
+  }
+  
+  if (remove_phase_outliers) {
+      phases_med <- stats::median(phases)
+      phases_mad <- stats::mad(phases)
+      upper_lim_phases <- phases_mad * phase_outlier_thresh
+      bad_phases <- Mod(as.numeric(phases - phases_med)) > upper_lim_phases
+  }
+  
+  if (remove_amp_outliers) {
+      amps_med <- stats::median(amps)
+      amps_mad <- stats::mad(amps)
+      upper_lim_amps <- amps_mad * amp_outlier_thresh
+      bad_amps <- Mod(as.numeric(amps - amps_med)) > upper_lim_amps
+  }
+  
   # apply to original data
   t_orig <- rep(seconds(mrs_data), each = Nspec(mrs_data))
   t_array <- array(t_orig, dim = dim(mrs_data$data))
@@ -109,7 +162,8 @@ rats <- function(mrs_data, ref = NULL, xlim = c(4, 0.5), max_shift = 20,
   if (!phase_corr) phases <- 0
   
   phase_array <- array(phases, dim = dim(mrs_data$data))
-  mod_array <- exp(2i * pi * t_array * shift_array + 1i * phase_array * pi / 180)
+  mod_array   <- exp(2i * pi * t_array * shift_array + 1i * phase_array * 
+                     pi / 180)
   mrs_data$data <- mrs_data$data * mod_array
  
   # maintain original intensities for bl_matched_spec and bl output
@@ -117,6 +171,10 @@ rats <- function(mrs_data, ref = NULL, xlim = c(4, 0.5), max_shift = 20,
     corr_spec <- scale_mrs_amp(corr_spec, 1 / amps)
     bl_spec   <- scale_mrs_amp(bl_spec,   1 / amps)
   }
+  
+  if (remove_amp_outliers) mrs_data <- mask_dyns(mrs_data, bad_amps)
+  if (remove_freq_outliers) mrs_data <- mask_dyns(mrs_data, bad_freqs)
+  if (remove_phase_outliers) mrs_data <- mask_dyns(mrs_data, bad_phases)
   
   # results
   res <- list(corrected = mrs_data, phases = -phases, shifts = -shifts,